remove experiment related code
