+++ Core/PixelEvent.swift
@@ -302,14 +302,7 @@ extension Pixel {
         case remoteMessageDismissed
         case remoteMessageShownPrimaryActionClicked
         case remoteMessageShownSecondaryActionClicked
-
-        // MARK: mac promo
-
-        case macPromoSheetShownUnique
-        case macPromoSheetDismissed
-        case macPromoPrimaryActionClicked
-        case shareLink
-
+        
         // MARK: debug pixels
         case dbCrashDetected
 
@@ -718,13 +711,6 @@ extension Pixel.Event {
         case .remoteMessageDismissed: return "m_remote_message_dismissed"
         case .remoteMessageShownPrimaryActionClicked: return "m_remote_message_primary_action_clicked"
         case .remoteMessageShownSecondaryActionClicked: return "m_remote_message_secondary_action_clicked"
-
-        // MARK: mac promo experiment
-
-        case .macPromoSheetShownUnique: return "m_macpromo_sheet_shown_unique"
-        case .macPromoSheetDismissed: return "m_macpromo_sheet_dismissed"
-        case .macPromoPrimaryActionClicked: return "m_macpromo_primary_action_clicked"
-        case .shareLink: return "m_share_link"
 
         // MARK: debug pixels
